requirements.txt
python-chess>=1.10.0
Chessnut>=2.4.0
rich>=13.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
